+++ packages/widgetbook_generator/CHANGELOG.md
@@ -3,6 +3,7 @@
  - **FEAT**: Add Dart 3 and Flutter 3.10 support. ([#676](https://github.com/widgetbook/widgetbook/pull/676))
  - **BREAKING**: Drop Flutter 2 support. ([#676](https://github.com/widgetbook/widgetbook/pull/676))
  - **BREAKING**: Remove code generation support for everything except `@App` and `@UseCase` that generate a list called `directories`. You should now configure Widgetbook's properties (i.e. `addons`, `appBuilder`, etc.) manually. ([#663](https://github.com/widgetbook/widgetbook/pull/663))
+ - **FIX**: Remove `<dynamic>` from Generic Widgets names. ([#700](https://github.com/widgetbook/widgetbook/pull/700))
 
 ## 3.0.0-beta.11